Fix force fully evolved for rival's starter (#164)
Because rival's starter is special as we keep it from start to finish,
it wasn't getting force_fully_evolved applied to it properly.

We cannot use `get_random_pokemon` as we need to make sure that we
don't evolve's rival starter randomly throughout the fights. Instead, if
we're past the force_fully_evolved threshold, we just resolve the max
evolution and set the pokemon as that.

An easy way to test this patch is to apply the following diff and then
run the fuzzer.

```diff
diff --git a/worlds/pokemon_crystal/rom.py b/worlds/pokemon_crystal/rom.py
index 9d9be0c3..4c117232 100644
--- a/worlds/pokemon_crystal/rom.py
+++ b/worlds/pokemon_crystal/rom.py
@@ -261,6 +261,9 @@ def generate_output(world: "PokemonCrystalWorld", output_directory: str, patch:
         address += trainer_data.name_length + 1  # skip name and type
         for pokemon in trainer_data.pokemon:
             pokemon_data = [pokemon.level, data.pokemon[pokemon.pokemon].id]
+            if world.options.randomize_trainer_parties and pokemon.level >= world.options.force_fully_evolved and not trainer_name == "LASS_3":
+                assert not data.pokemon[pokemon.pokemon].evolutions, f"{pokemon.level}, {pokemon.pokemon}, {trainer_name}"
             if pokemon.item is not None:
                 item_id = item_const_name_to_id(pokemon.item)
                 pokemon_data.append(item_id)
```

---------

Co-authored-by: cheerioschelsea <37007731+cheerioschelsea@users.noreply.github.com>