Avoid some useless copies (#215)
* Avoid a useless copy in get_random_types

This is ~10 times faster than the original. Not that it might be
tempting to use `random.sample(data, type_numbers)` but that's actually
slower.

results for a million calls:

```
old
4.820971396751702

new with sample
0.8043631492182612

new
0.5868077166378498
```

* Remove 2 useless copies of the crystal type data

I have no idea why it was copied in the first place, nothing in here
modifies the original value